standardize function access, allow setting to None
Standardize the way functions are accessed.  There were a few locations
with a try / except AttributeError, while most used the three argument
form of getattr.  Now all use the latter method, with the third argument
as None, followed by an explicit check for None.  This allows functions
to be set to None to indicate that nothing should be done.