Terminal.SGR: fix '0' (clear_all)
Previously, the clear_all() function wtihin Terminal.SGR() returned
a function rather than doing something useful.  I blindly copied the
format from set_attr() and clear_attr(), which both get called in the
construction of the dictionary.  However, clear_all is simply a value in
the dictionary, not called, so it should do the actual of clearing the
attributes.  This is now fixed.