fix REP by properly tracking previous character
Previously, the previous character was simply the previous character
input, which does not work correctly for REP ('b' would always be
produced, since that is the character for the REP command, CSI b).  Now,
I keep track of the current and previous character processed in the
ground state.  This seems to work OK in testing when parsing the
'CSI Pt b' sequence, but it probably won't work if you call REP()
directly.