fix BS past EOL; add clip_column()
BS must act like IND in that, if past end of the line, the cursor must
be repositioned to the last position before moving.  Since this is now
used in two places, and I imagine more will follow, I created a new
function, clip_column(), to do this for us.