rawcat: add -b and -w options
Add --buffer-size (-b) option, which specifies the number of bytes to
read at a time.

Add --wait (-w) option, which causes the program to wait this many
milliseconds after each buffer.  If given and -b is not, the buffer size
defaults to 1.

Update the manual page of vt100.py to mention this new feature.