treat private parameters as intermediates
Technically, starting a control sequence with one of '<=>?' causes the
parameter to be parsed specially, but the command is the same.  In
practice, it is used as an alternative to the intermediate byte, and is
used to introduce new (or modified) commands.  Therefore, it is easiest
if we just treat it as part of the command sequence, rather than making
each function check for this special value up front.