add --geometry=detect option
Add a new choice for geometry: "detect".  If given, the current
terminal's geometry is detected by running `stty size`, which returns
the rows and columns as integers, separated by a space.  However, this
only works on Linux.

On Solaris, it is possible to run `stty -a` and parse the output, but
this is a pain.  Unless there is a Solaris user out there, I'm not going
to bother.

It is also possible to use ioctl or termcap, but this doesn't seem worth
the effort.