clean up implementation of SimpleConfigParser
Instead of adding my own methods to the configuration parser, instead
modify the low-level _read() and high-level get() functions to allow for
a default section.

Add an keyword-only argument, 'default_section', to the constructor,
which allows the caller to set the name of the default section.  This
way, if given something sane like "main", the end-user can get back to
the main section within a config file.