switch to class-based output formatters
Instead of passing in a line-formatting function to Terminal and
printing the pre and post in main(), instead pass in a formatter object
that has a format() method that automatically adds a pre- and post-
outputs.

This is a more sane interface, groups all the HTML stuff together, and
will make it easier in the future to add options to the formatter.