implement inverse attribute
There is no HTML "inverse" style, so we have to fake it by explicitly
swapping foreground and background colors.  However, this is a problem
if the character does not have explicit foreground and background
colors, and if the <pre> tag's foreground and background were also not
set.  So, we will introduce a new option to give the fallback colors in
that case.